removed duplicate in progress component, combined into single
