--- src/components/pages/home/inprogress/in-progress-collection.stories.mdx
@@ -1,95 +1,0 @@
-import {Meta, Story, Canvas} from '@storybook/addon-docs/blocks'
-import InProgressCollection from './in-progress-collection'
-
-export const collection = {
-  title: 'Develop Basic Web Apps with Vue.js',
-  lessons: [
-    {
-      title: 'Create a Basic Component using Vue.js',
-      slug: 'vue-js-create-a-basic-component-using-vue-js',
-      path: '/lessons/vue-js-create-a-basic-component-using-vue-js',
-    },
-    {
-      title: 'Create a List Component in Vue.js',
-      slug: 'vue-js-create-a-list-component-in-vue-js',
-      path: '/lessons/vue-js-create-a-list-component-in-vue-js',
-    },
-    {
-      title: 'Create Filters in Vue.js',
-      slug: 'vue-js-create-filters-in-vue-js',
-      path: '/lessons/vue-js-create-filters-in-vue-js',
-    },
-    {
-      title: 'Use Vue.js Component Computed Properties',
-      slug: 'vue-js-use-vue-js-component-computed-properties',
-      path: '/lessons/vue-js-use-vue-js-component-computed-properties',
-    },
-    {
-      title: 'Use Vue.js Watchers to Respond to Async Updates',
-      slug: 'vue-js-use-vue-js-watchers-to-respond-to-async-updates',
-      path: '/lessons/vue-js-use-vue-js-watchers-to-respond-to-async-updates',
-    },
-    {
-      title: 'Update Attributes, Classes and Styles in Vue.js with v-bind',
-      slug: 'vue-js-update-attributes-classes-and-styles-in-vue-js-with-v-bind',
-      path:
-        '/lessons/vue-js-update-attributes-classes-and-styles-in-vue-js-with-v-bind',
-    },
-    {
-      title: 'Conditionally Render DOM Elements in Vue.js ',
-      slug: 'vue-js-conditionally-render-dom-elements-in-vue-js',
-      path: '/lessons/vue-js-conditionally-render-dom-elements-in-vue-js',
-    },
-    {
-      title: 'Handle DOM Events in Vue.js with v-on',
-      slug: 'vue-js-handle-dom-events-in-vue-js-with-v-on',
-      path: '/lessons/vue-js-handle-dom-events-in-vue-js-with-v-on',
-    },
-    {
-      title: 'Bind Values to Input Fields in Vue.js',
-      slug: 'vue-js-bind-values-to-input-fields-in-vue-js',
-      path: '/lessons/vue-js-bind-values-to-input-fields-in-vue-js',
-    },
-    {
-      title: 'Build Reusable Vue.js Components',
-      slug: 'vue-js-build-reusable-vue-js-components',
-      path: '/lessons/vue-js-build-reusable-vue-js-components',
-    },
-    {
-      title: 'Dynamic Vue.js Components with the component element',
-      slug: 'vue-js-dynamic-vue-js-components-with-the-component-element',
-      path:
-        '/lessons/vue-js-dynamic-vue-js-components-with-the-component-element',
-    },
-  ],
-  square_cover_480_url:
-    'https://d2eip9sf3oo6c2.cloudfront.net/series/square_covers/000/000/083/square_480/EGH_VueJS_Final.png',
-  type: 'course',
-  path: '/courses/develop-basic-web-apps-with-vue-js',
-  progress: {
-    lesson_count: 11,
-    completed_lesson_count: 1,
-    is_completed: false,
-    completed_at: null,
-    completed_lessons: [
-      {
-        title: 'Use Vue.js Component Computed Properties',
-        slug: 'vue-js-use-vue-js-component-computed-properties',
-        path: '/lessons/vue-js-use-vue-js-component-computed-properties',
-      },
-    ],
-  },
-}
-
-<Meta title="Components/Home/In Progress" />
-
-# Featured Resource
-
-<Canvas>
-  <Story name="Progress Cards">
-    <div className="grid md:grid-cols-2 gap-4 bg-gray-50 p-10">
-      <InProgressCollection collection={collection} />
-      <InProgressCollection small collection={collection} />
-    </div>
-  </Story>
-</Canvas>
