+++ src/components/pages/users/dashboard/activity/recently-watched-section.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import {FunctionComponent} from 'react'
-import ResourceCard from './resource'
+import InProgressResource from './inprogress-resource'
 import {Resource} from 'types'
 import {filter, uniq} from 'lodash'
 
@@ -27,7 +27,7 @@ const RecentlyWatched: FunctionComponent<RecentlyWatchedProps> = ({
         </h2>
         <div className="space-y-3">
           {lessonsCompleted.map((resource: any) => {
-            return <ResourceCard {...resource} key={resource.id} />
+            return <InProgressResource resource={resource} key={resource.id} />
           })}
         </div>
       </div>
